Remove implementation code from header files
* implementation code of nanovg_gl.h has been moved to nanovg_gl.c
* nvgCreateGL has been moved to nanovg.c
* which OpenGL version to use has become a compile time switch
* premake4 has been updated accordingly
* examples have been updated

Having implemenation code split between header files and the .c files is
a bad idea.
First of all it complicates the build process and the usage/development of the
library.
Secondly, if the library is dynamically linked this could lead
to severe problems as the implementation code in the header files might be a different
version than the code in the library (.c files).
