+++ example/example_gl3.c
@@ -17,16 +17,9 @@
 //
 
 #include <stdio.h>
-#ifdef NANOVG_GLEW
-#	include <GL/glew.h>
-#endif
-#ifdef __APPLE__
-#	define GLFW_INCLUDE_GLCOREARB
-#endif
+#define GLFW_INCLUDE_NONE
 #include <GLFW/glfw3.h>
 #include "nanovg.h"
-#define NANOVG_GL3_IMPLEMENTATION
-#include "nanovg_gl.h"
 #include "demo.h"
 #include "perf.h"
 
@@ -94,7 +87,6 @@ int main()
 	glfwSetKeyCallback(window, key);
 
 	glfwMakeContextCurrent(window);
-#ifdef NANOVG_GLEW
 	glewExperimental = GL_TRUE;
 	if(glewInit() != GLEW_OK) {
 		printf("Could not init glew.\n");
@@ -102,12 +94,11 @@ int main()
 	}
 	// GLEW generates GL error because it calls glGetString(GL_EXTENSIONS), we'll consume it here.
 	glGetError();
-#endif
 
 #ifdef DEMO_MSAA
-	vg = nvgCreateGL3(NVG_STENCIL_STROKES | NVG_DEBUG);
+	vg = nvgCreateGL(NVG_STENCIL_STROKES | NVG_DEBUG);
 #else
-	vg = nvgCreateGL3(NVG_ANTIALIAS | NVG_STENCIL_STROKES | NVG_DEBUG);
+	vg = nvgCreateGL(NVG_ANTIALIAS | NVG_STENCIL_STROKES | NVG_DEBUG);
 #endif
 	if (vg == NULL) {
 		printf("Could not init nanovg.\n");
@@ -186,7 +177,7 @@ int main()
 
 	freeDemoData(vg, &data);
 
-	nvgDeleteGL3(vg);
+	nvgDeleteGL(vg);
 
 	printf("Average Frame Time: %.2f ms\n", getGraphAverage(&fps) * 1000.0f);
 	printf("          CPU Time: %.2f ms\n", getGraphAverage(&cpuGraph) * 1000.0f);
